refactor: move the incident summary translation onto its read model
`IncidentReportSummaryMapper` was the last module in the mappers directory.
Unlike the two just removed it does real work — it drops `reporter_user_id`
and stringifies UUIDs — so the translation is preserved verbatim as
`IncidentReportSummary.from_report/1`, on the struct that already documents
why the reporter FK is absent.

Folds the mapper's four near-identical FK tests into one tabular test and
merges the file into the existing read-model test, adding a case pinning the
dropped field.

The mappers directory is now empty and removed; no `to_read_model/1` or
`from_schema/1` remains anywhere in lib.
